+++ CHANGELOG.md
@@ -125,6 +125,8 @@ is used for timestamps for synthesized events.
 
 ### Additions
 
+- Add `WmWindowProperty::Other` to forward unrecognized X11 property changes to the compositor.
+
 - ExtBackgroundEffect protocol is now available in `smithay::wayland::background_effect` module.
 
 `crate::input::dnd` was introduced to enable implementation of Drag&Drop operations on custom types.